separar_pareceres/requirements.txt
altgraph==0.17.2
deskew==1.0.26
future==0.18.2
imageio==2.19.3
networkx==2.8.4
numpy==1.23.0
opencv-python==4.6.0.66
packaging==21.3
pdf2image==1.16.0
pefile==2022.5.30
Pillow==9.1.1
pycpdflib==2.5
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.7
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pytesseract==0.3.9
PyWavelets==1.3.0
pywin32-ctypes==0.2.0
regex==2022.6.2
scikit-image==0.19.3
scipy==1.8.1
tifffile==2022.5.4
